requirements.txt
absl-py==1.0.0
asttokens==2.0.5
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.1
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fonttools==4.31.2
google-auth==2.6.0
google-auth-oauthlib==0.4.6
graphviz==0.19.1
grpcio==1.44.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.2
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
kiwisolver==1.4.2
LibMTL==1.1.5
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
networkx==2.8.4
nltk==3.7
numpy==1.20.0
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
pyzmq==23.2.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
stack-data==0.3.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2022.5.4
torch==1.8.0
torchsummary==1.5.1
torchvision==0.9.0
torchviz==0.0.2
tornado==6.1
tqdm==4.63.1
traitlets==5.3.0
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
zipp==3.7.0
